LayoutTests/fast/forms/date-multiple-fields/date-multiple-fields-keyboard-events.html

<!DOCTYPE html>
<html>
<head>
<script src="../../js/resources/js-test-pre.js"></script>
</head>
<body>
<p id="description"></p>
<p>
Please run this with DumpRenderTree.
</p>
Test following keys:
<ul>
    <li>Digits</li>
    <li>Left/Right - Move focus field inside element</li>
    <li>Up/Down - Increment/decrement value of focus field</li>
    <li>Tab - Move focus field</li>
    <li>Backspace - Make value empty</li>
</ul>
<input id="before">
<input id="input" type="date">
<input id="after">
<div id="console"></div>
<script>
description('Multiple fields UI of month input type with keyboard events');
if (window.internals)
    internals.settings.setLangAttributeAwareFormControlUIEnabled(true);
var input = document.getElementById('input');

function keyDown(key, modifiers)
{
    if (!window.eventSender)
        return;
    eventSender.keyDown(key, modifiers);
}

function beginTest(title, opt_value, opt_min, opt_max)
{
    debug('== ' + title + ' ==');
    input.value = opt_value || '';
    input.min = opt_min ? opt_min : '';
    input.max = opt_max ? opt_max : '';
    input.blur();
    input.focus();
}

// We assume the date format is MM/dd/yyyy.

beginTest('Digit keys');
keyDown('9'); // -> 09/[dd]/yyyy
keyDown('2'); // -> 09/[02]/yyyy
keyDown('0'); // -> 09/20/[yyyy]
keyDown('1'); // -> 09/20/[0001]
keyDown('2'); // -> 09/20/[0012]
keyDown('A'); // Ignored.
shouldBeEqualToString('input.value', '0012-09-20');

beginTest('Digit keys starting with zero');
keyDown('0'); // -> [00]/dd/yyyy
keyDown('2'); // -> 02/[dd]/yyyy
keyDown('0'); // -> 02/[00]/yyyy
keyDown('3'); // -> 02/03/[yyyy]
keyDown('0'); // -> 02/03/[0000]
keyDown('0'); // -> 02/03/[0000]
keyDown('0'); // -> 02/03/[0000]
keyDown('4'); // -> 02/03/[0004]
keyDown('4'); // -> 02/03/[0044]
shouldBeEqualToString('input.value', '0044-02-03');

beginTest('Digit keys and backspace key');
keyDown('1'); // -> [01]/dd/yyyy
keyDown("\b"); // -> [mm]/20/2012
keyDown('5'); // -> 05/[dd]/yyyy
keyDown('6'); // -> 05/06/[yyyy]
keyDown("\b"); // -> 05/06/[yyyy]
keyDown('7'); // -> 05/06/[0007]
keyDown("\b"); // -> 05/06/[yyyy]
keyDown('8'); // -> 05/06/[0008]
shouldBeEqualToString('input.value', '0008-05-06');

// FIXME: We should test type ahead time out. When event.leapForward() affects
// keyboard event time stamp, we can uncomment this fragment.
/*
beginTest('Digit keys with type ahead timeout');
keyDown('1'); // -> [01]/dd/yyyy
leapForward(1100); // -> 01/[dd]/yyyy
keyDown('1'); // -> 01/[01]/yyyy
keyDown('5'); // -> 01/15/[yyyy]
keyDown('6'); // -> 01/15/[0006]
shouldBeEqualToString('input.value', '0006-01-15');
*/

beginTest('Left/Right keys', '2012-09-29');
keyDown('rightArrow'); // -> 09/[29]/2012
keyDown('5'); // -> 09/05/[2012]
keyDown('leftArrow'); // -> 09/[05]/2012
keyDown('6'); // -> 09/06/[2012]
shouldBeEqualToString('input.value', '2012-09-06');
keyDown('leftArrow'); // -> 09/[06]/2012
keyDown('leftArrow'); // -> [09]/06/2012
keyDown('leftArrow'); // -> [09]/06/2012
shouldBeEqualToString('document.activeElement.id', 'input');

beginTest('Up/Down keys', '2012-09-29');
keyDown('upArrow'); // -> [10]/29/2012
shouldBeEqualToString('input.value', '2012-10-29');
keyDown('downArrow'); // -> [09]/29/2012
keyDown('downArrow'); // -> [08]/29/2012
shouldBeEqualToString('input.value', '2012-08-29');

beginTest('Up/Down keys on empty value');
keyDown('downArrow'); //  -> [12]/dd/yyyy
keyDown('downArrow'); //  -> [11]/dd/yyyy
keyDown('rightArrow'); // -> 11/[dd]/yyyy
keyDown('upArrow'); //    -> 11/[01]/yyyy
keyDown('upArrow'); //    -> 11/[02]/yyyy
keyDown('downArrow'); //  -> 11/[01]/yyyy
keyDown('rightArrow'); // -> 11/01/[yyyy]
var currentYear = new Date().getFullYear();
keyDown('upArrow'); // -> 11/01/[current year]
shouldBe('input.value', 'currentYear + "-11-01"');

beginTest('Up/Down keys on empty value 2');
keyDown('upArrow'); //    -> [01]/dd/yyyy
keyDown('upArrow'); //    -> [02]/dd/yyyy
keyDown('rightArrow'); // -> 02/[dd]/yyyy
keyDown('downArrow'); //  -> 02/[31]/yyyy
keyDown('downArrow'); //  -> 02/[30]/yyyy
keyDown('downArrow'); //  -> 02/[29]/yyyy
keyDown('downArrow'); //  -> 02/[28]/yyyy
keyDown('rightArrow'); // -> 02/28/[yyyy]
currentYear = new Date().getFullYear();
keyDown('downArrow'); // -> 02/28/[current year]
shouldBe('input.value', 'currentYear + "-02-28"');

beginTest('Tab key', '2012-09-30');
keyDown('\t'); // -> 09/[30]/2012
keyDown('5'); // -> 09/05/[2012]
shouldBeEqualToString('input.value', '2012-09-05');
keyDown('\t', ['shiftKey']); // -> 09/[05]/2012
keyDown('7'); // -> 09/07/[2012]
shouldBeEqualToString('input.value', '2012-09-07');
keyDown('\t'); // -> Focus out.
shouldBeEqualToString('document.activeElement.id', 'after');

beginTest('Shfit+Tab key', '2012-09-30');
after.focus();
keyDown('\t', ['shiftKey']); // -> 09/30/[yyyy]
keyDown('3'); // -> 09/30/[0003]
shouldBeEqualToString('input.value', '0003-09-30');
keyDown('\t', ['shiftKey']); // -> 09/[30]/0003
keyDown('\t', ['shiftKey']); // -> [09]/30/0003
keyDown('\t', ['shiftKey']); // -> Focus out.
shouldBeEqualToString('document.activeElement.id', 'before');

beginTest('Up key on maximum value', '275760-09-13');
keyDown('upArrow'); // -> [10]/13/275760
keyDown('\t'); //      -> 10/[13]/275760
keyDown('upArrow'); // -> 10/[14]/275760
keyDown('\t'); //      -> 10/14/[275760]
keyDown('upArrow'); // -> 10/14/[0001]
shouldBeEqualToString('input.value', '0001-10-14');
beginTest('Up key with a maximum attribute', '1999-12-31', '1000-01-01', '1999-12-31');
keyDown('upArrow'); // -> [01]/31/1999
keyDown('\t'); //      -> 01/[31]/1999
keyDown('upArrow'); // -> 01/[01]/1999
keyDown('\t'); //      -> 01/01/[1999]
keyDown('upArrow'); // -> 01/01/[1000]
shouldBeEqualToString('input.value', '1000-01-01');

beginTest('Down key on minimum value', '0001-01-01', 'bad min', 'wrong max');
keyDown('downArrow'); // -> [12]/01/0001
keyDown('\t'); //        -> 12/[01]/0001
keyDown('downArrow'); // -> 12/[31]/0001
keyDown('\t'); //        -> 12/31/[0001]
keyDown('downArrow'); // -> 12/31/[275760], which is greater than the hard limit.
shouldBeEqualToString('input.value', '');
beginTest('Down key with a minimum attribute', '1000-01-01', '1000-01-01');
keyDown('downArrow'); // -> [12]/01/1000
keyDown('\t'); //        -> 12/[01]/1000
keyDown('downArrow'); // -> 12/[31]/1000
keyDown('\t'); //        -> 12/31/[1000]
keyDown('downArrow'); // -> 12/31/275760, which is greater than the hard limit.
shouldBeEqualToString('input.value', '');

beginTest('Inconsistent min-max attributes', '1999-12-31', '1999-12-31', '1000-01-01');
keyDown('\t'); // -> 12/[31]/1999
keyDown('\t'); // -> 12/31/[1999]
keyDown('upArrow'); // -> 12/31/[1000].  1000 is the swapped minimum year.
shouldBeEqualToString('input.value', '1000-12-31');
keyDown('downArrow'); // -> 12/31/[1999]
shouldBeEqualToString('input.value', '1999-12-31');

beginTest('Make an invalid date', '2012-02-01');
keyDown('\t'); //        -> 02/[01]/2012
keyDown('downArrow'); // -> 02/[31]/2012
shouldBeEqualToString('input.value', ''); // 2012-02-31 is not a valid date.

beginTest('Backspace key', '2012-09-20');
keyDown("\b"); // -> [mm]/20/2012
shouldBeEqualToString('input.value', '');

beginTest('Delete key', '2012-09-30');
keyDown("delete"); // -> [mm]/30/2012
shouldBeEqualToString('input.value', '');

beginTest('Typeahead', '2012-12-31');
keyDown('rightArrow'); // -> 12/[31]/2012
keyDown('1'); //          -> 12/[01]/2012
shouldBeEqualToString('input.value', '2012-12-01');
keyDown('leftArrow'); //  -> [12]/01/2012
keyDown('rightArrow'); // -> 12/[01]/2012
keyDown('2'); //          -> 12/[02]/2012
shouldBeEqualToString('input.value', '2012-12-02');

input.setAttribute("lang", "he-il");
beginTest('RTL focus navigation', '2012-09-28');
debug('The tests in this block fail on platforms without the lang-attribute-aware-form-control-UI feature.');
// Both of the logical order and visual order are: dd/MM/yyyy
// Initial state:            [28]/09/2012
keyDown('upArrow'); //    -> [29]/09/2012
keyDown('rightArrow'); // -> [01]/09/2012
keyDown('1'); //          -> [01]/09/2012
shouldBeEqualToString('input.value', '2012-09-01');
keyDown('\t'); //         -> 01/[09]/2012
keyDown('2'); //          -> 01/02/[2012]
shouldBeEqualToString('input.value', '2012-02-01');
keyDown('\t', ['shiftKey']); // -> 01/[02]/2012/
keyDown('3'); //          -> 01/03/[2012]
shouldBeEqualToString('input.value', '2012-03-01');
input.removeAttribute("lang");

beginTest('Disabled/readonly', '2012-10-08');
input.disabled = true;
keyDown('upArrow'); // 10/08/2012
shouldBeEqualToString('input.value', '2012-10-08');
input.disabled = false;
input.focus();
keyDown('upArrow'); // [11]/08/2012
shouldBeEqualToString('input.value', '2012-11-08');
input.readOnly = true;
keyDown('upArrow'); // 11/08/2012
shouldBeEqualToString('input.value', '2012-11-08');
input.readOnly = false;
input.focus();
keyDown('upArrow'); // [12]/08/2012
shouldBeEqualToString('input.value', '2012-12-08');

</script>
<script src="../../js/resources/js-test-post.js"></script>
</body>
</html>
